+++ osu-trainer.sln
@@ -3,18 +3,18 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29911.84
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "osu-trainer", "osu! BPM Changer\osu-trainer.csproj", "{88026F3A-2BC2-4542-87CE-6A7E796FD2DD}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "osu-trainer", "osu-trainer\osu-trainer.csproj", "{88026F3A-2BC2-4542-87CE-6A7E796FD2DD}"
 	ProjectSection(ProjectDependencies) = postProject
 		{D117800F-072D-4AE4-9679-3E2A129A1A3C} = {D117800F-072D-4AE4-9679-3E2A129A1A3C}
 		{9960D641-300A-432C-AA04-C351A99B9ADB} = {9960D641-300A-432C-AA04-C351A99B9ADB}
 		{7D292085-3F6F-4299-B941-4D2C185DA37E} = {7D292085-3F6F-4299-B941-4D2C185DA37E}
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "osu!BMAPI", "osu! BPM Changer\BMAPI\osu!BMAPI.csproj", "{7D292085-3F6F-4299-B941-4D2C185DA37E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "osu!BMAPI", "osu-trainer\BMAPI\osu!BMAPI.csproj", "{7D292085-3F6F-4299-B941-4D2C185DA37E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OsuMemoryDataProvider", "osu! BPM Changer\OsuMemoryDataProvider\OsuMemoryDataProvider.csproj", "{D117800F-072D-4AE4-9679-3E2A129A1A3C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OsuMemoryDataProvider", "osu-trainer\OsuMemoryDataProvider\OsuMemoryDataProvider.csproj", "{D117800F-072D-4AE4-9679-3E2A129A1A3C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProcessMemoryDataFinder", "osu! BPM Changer\ProcessMemoryDataFinder\ProcessMemoryDataFinder.csproj", "{9960D641-300A-432C-AA04-C351A99B9ADB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProcessMemoryDataFinder", "osu-trainer\ProcessMemoryDataFinder\ProcessMemoryDataFinder.csproj", "{9960D641-300A-432C-AA04-C351A99B9ADB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -57,6 +57,7 @@ Global
 		{D117800F-072D-4AE4-9679-3E2A129A1A3C}.Debug|x86.ActiveCfg = Debug|x86
 		{D117800F-072D-4AE4-9679-3E2A129A1A3C}.Debug|x86.Build.0 = Debug|x86
 		{D117800F-072D-4AE4-9679-3E2A129A1A3C}.Release|Any CPU.ActiveCfg = Release|x86
+		{D117800F-072D-4AE4-9679-3E2A129A1A3C}.Release|Any CPU.Build.0 = Release|x86
 		{D117800F-072D-4AE4-9679-3E2A129A1A3C}.Release|x64.ActiveCfg = Release|x64
 		{D117800F-072D-4AE4-9679-3E2A129A1A3C}.Release|x64.Build.0 = Release|x64
 		{D117800F-072D-4AE4-9679-3E2A129A1A3C}.Release|x86.ActiveCfg = Release|x86
@@ -68,6 +69,7 @@ Global
 		{9960D641-300A-432C-AA04-C351A99B9ADB}.Debug|x86.ActiveCfg = Debug|x86
 		{9960D641-300A-432C-AA04-C351A99B9ADB}.Debug|x86.Build.0 = Debug|x86
 		{9960D641-300A-432C-AA04-C351A99B9ADB}.Release|Any CPU.ActiveCfg = Release|x86
+		{9960D641-300A-432C-AA04-C351A99B9ADB}.Release|Any CPU.Build.0 = Release|x86
 		{9960D641-300A-432C-AA04-C351A99B9ADB}.Release|x64.ActiveCfg = Release|x64
 		{9960D641-300A-432C-AA04-C351A99B9ADB}.Release|x64.Build.0 = Release|x64
 		{9960D641-300A-432C-AA04-C351A99B9ADB}.Release|x86.ActiveCfg = Release|x86
