+++ osu-trainer/DifficultyCalculator.cs
