+++ osu-trainer/OsuMemoryDataProvider/IOsuMemoryReader.cs
