+++ osu-trainer/OsuMemoryDataProvider/OsuMemoryReader.cs
