+++ osu-trainer/OsuMemoryDataProvider/OsuMemoryStatus.cs
