+++ osu-trainer/OsuMemoryDataProvider/Properties/AssemblyInfo.cs
