+++ osu-trainer/OsuMemoryDataProvider/SignatureNames.cs
