+++ osu-trainer/ProcessMemoryDataFinder/API/Internals.cs
