+++ osu-trainer/ProcessMemoryDataFinder/API/MemoryReader.cs
