+++ osu-trainer/ProcessMemoryDataFinder/API/MemoryReaderEx.cs
