+++ osu-trainer/ProcessMemoryDataFinder/API/ProcessMemoryReader.cs
