+++ osu-trainer/ProcessMemoryDataFinder/API/Sig.cs
