+++ osu-trainer/ProcessMemoryDataFinder/API/SigScan.cs
