+++ osu-trainer/ProcessMemoryDataFinder/ProcessExtensions.cs
