+++ osu-trainer/ProcessMemoryDataFinder/Properties/AssemblyInfo.cs
