+++ osu-trainer/osu-trainer.csproj
